requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argbind==0.3.9
assemblyai==0.41.5
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
build==1.2.2.post1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
dac==0.4.3
decorator==5.2.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
docstring_parser==0.16
edge-tts==7.0.2
einops==0.8.1
executing==2.2.0
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
flatten-dict==0.4.2
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
grpcio==1.73.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
importlib_resources==6.5.2
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.19.6
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyloudnorm==0.1.1
pyparsing==3.2.3
pypiwin32==223
pyproject_hooks==1.2.0
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyttsx3==2.98
pywin32==310
PyYAML==6.0.2
randomname==0.2.1
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
srt==3.5.3
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
toml==0.10.2
torch==2.5.1
torch-stoi==0.2.3
torchaudio==2.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
